Automatic RHMP integration (#533)
* Enable RHMP automatically

RHMP will be enabled automatically when MeterDefinition CRD is detected and the feature wasn't set manually.

* Log automatic RHMP detection

When spec.rhmpEnabled wasn't specified we want to log the outcome of the automatic detection

* Add RHMP status next to pod statuses

This allows for viewing the RHMP status. It's especially useful when it's been set automatically (spec.rhmpEnabled not present).

Co-authored-by: kacperkrzyzak <65035142+kacperkrzyzak@users.noreply.github.com>